+++ CHANGES.md
@@ -5,6 +5,8 @@
 - check whether file exists before reading it
 - implement the ability to require 2FA
 - switch to `Tpl`, completely remove Twig
+- remove `tls-auth` and `enableCompression` from profile config, no longer 
+  supported
 
 ## 1.3.2 (2018-12-05)
 - remove PHP error suppression